Makefile tweaks...
Changing the all: target to ALL=all and ${ALL}: for increased
flexibility in subdir Makefiles.  Also making ${ALL}: ${DEPENDS}
adds the opportunity to add dependencies in subdir Makefiles
for things like generating U-Boot image files for example as part
of the all: target.

Done in such a way that QtCreator *should* still be able to use
this makefile build system completely unchanged.  But that's yet to be tested.