Progress! I know how to boot this image file...
U-Boot is a bit perplexing at times.  Support for FIT image files
is still a bit of a mystery, BUT...

I know how to build a FIT image file that can be retrieved with TFTP and can
boot linux, but it takes some shenanigans on the target
using imxtract.  More on this on the wiki for this project.