Booting via TFTP with a script...
Making progress.  Instead of using U-Boot FIT file which contains
the kernel, the FDT image, and the initrd image, use a script to
fetch those from the TFTP server with individual "tftp" commands.
Placing all of those things in a script file along with a "bootz"
command and a kernel command line arguments environment variable
means we get to see the kernel messages as it's booting.